Add different hands and and gripperjoint
