+++ src/reachy2_sdk/grippers/__init__.py
@@ -1,0 +1,5 @@
+"""Grippers module.
+
+This module provides Reachy's grippers.
+
+"""
